Update confirmation plugin to remove data-*
To reuse as much as possible, the configuration should be done upon
instantiation of the plugin rather than with data attributes. This
is now possible. The only data attribute respected is data-title
and will be used when in configuration the title is not set.

Documentation is updated and to allow easier readability, the plugin
is renamed to confirm.